ex04 (Missing tests)

+++ modulo2/ex04/asm.h
@@ -1,4 +1,4 @@
 #ifndef ASM_H
 #define ASM_H
-int sum_v3(void);
+long sum_v3(void);
 #endif
